requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.1+cpu
numpy<2.0.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
sqlalchemy==2.0.25
neo4j==5.17.0
spacy==3.6.1
sentence-transformers==2.2.2
PyMuPDF==1.23.8
trafilatura==1.6.4
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
groq>=1.1.1
pymongo>=4.6.0
